extract cache features to own file, add reset function
